externalisation et ajout d'un id aléatoire
