+++ types/types.d.ts
@@ -1,0 +1,12 @@
+export type Contact = {
+  id: string;
+  name: string;
+  firstname: string;
+  email: string;
+  phone: string;
+};
+
+export interface CreateContactFormProps {
+  open: boolean;
+  setOpen: (open: boolean) => void;
+}
